src/: rm caching and pure-macaronic features.
We're focusing the "code blocks" flow instead, so let's remove all the
code we no longer need.
