--- src/maccarone/caching.mn.json
@@ -1,6 +1,0 @@
-{
-    "tagged_input_to_tagged_output": {
-        "inputs_hash": "d53ed422ea0be09307808bc7432781647d510717265044ad2d3ce37834c14bd9",
-        "value": "<completed id=\"0\">\ntry:\n    with open(self.path, \"r\") as f:\n        return json.load(f)\nexcept (FileNotFoundError, json.JSONDecodeError):\n    return {}\n</>\n\n<completed id=\"1\">\nwith open(self.path, \"w\") as f:\n    json.dump(cache, f, indent=4, sort_keys=True)\n</>"
-    }
-}
