--- src/maccarone/loader.mn.json
@@ -1,6 +1,0 @@
-{
-    "tagged_input_to_tagged_output": {
-        "inputs_hash": "b13a352b12549da61916187030713eb300e9b64bc3ea8df43de2faefa07c8171",
-        "value": "<completed id=\"0\">\n# No code needs to be written for this task.\n</>"
-    }
-}
