+++ src/maccarone/openai.py
@@ -11,11 +11,6 @@ import openai
 
 from openai import ChatCompletion
 
-from maccarone.caching import (
-    DiskCache,
-    CacheKeyMissingError,
-)
-
 logger = logging.getLogger(__name__)
 
 openai.api_key = os.getenv("OPENAI_API_KEY")
@@ -61,21 +56,11 @@ def complete_chat(
 
     return completion
 
-class CachedChatAPI:
-    def __init__(self, cache_path: str):
-        self.cache = DiskCache(cache_path)
-
+class ChatAPI:
     def complete_chat(
             self,
             chat_name: str,
             messages: list[dict[str, str]],
             model="gpt-4",
         ) -> str:
-        try:
-            return self.cache.get(chat_name, messages)
-        except CacheKeyMissingError:
-            completion = complete_chat(messages, model=model)
-
-            self.cache.set(chat_name, messages, completion)
-
-            return completion
+        return complete_chat(messages, model=model)
